+++ macro-magic/tests/test_item.rs
@@ -1,3 +1,4 @@
+#![recursion_limit = "150"]
 use ::macro_magic::*;
 
 __declare_features!(